new data model in transactionlist

+++ budgetiser/lib/shared/services/accountItem/accountItem.dart
@@ -63,7 +63,7 @@ class AccountItem extends StatelessWidget {
                     ),
                   ],
                 ),
-                BalanceText(balance),
+                BalanceText(balance.toDouble()),
               ],
             ),
           ),
